api/requirements.txt
flask==2.3.3
flask-restful==0.3.10
flask-cors==4.0.0
psycopg2-binary==2.9.7
sqlalchemy==2.0.21
gunicorn==21.2.0
python-dotenv==1.0.0
pyjwt==2.8.0 